// src/app/client/models/dto/insured-save-input.dto.ts
export class InsuredSaveInputDto {
    insuredName: string;
    insuredDocument: string;
    insuredAddressRoute: string;
    insuredAddressNumber: string;
    insuredAddressNeighborhood: string;
    insuredAddressCity: string;
    insuredAddressState: string;
    insuredAddressZipCode: string;
    insuredPhone: string;
    insuredEmail: string;
}